+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleCredentialModes.scala
@@ -28,7 +28,7 @@ object GoogleCredentialModes {
     * Gets a GoogleCredential from a pem file.
     */
   case class Pem(serviceAccountClientId: WorkbenchEmail, pemFile: File, serviceAccountUser: Option[WorkbenchEmail] = None) extends GoogleCredentialMode {
-    def toGoogleCredential(scopes: Seq[String]) = {
+    def toGoogleCredential(scopes: Seq[String]): GoogleCredential = {
       new GoogleCredential.Builder()
         .setTransport(httpTransport)
         .setJsonFactory(jsonFactory)